liboverlay: Fix the horizontal displacement issue on 3D video on 2D TV
In 3D video capture, right and left cameras are placed at different
view point. so there will be a difference in right and left views.
currently we are displaying left view on the primary panel and right
view on the external display.

Change: Display left view of 3D video on both primary and external
2D display devices.

CRs-Fixed: 368673

Change-Id: Id7917448560c915e053e7193fec990d978504db3