Display: Add support for dynamic framebuffer scaling
Add support for dynamically changing the framebuffer resolution.
To enable dynamic framebuffer resolution, the debug.fb.width and
debug.fb.height properties need to be set. If either property is
set to 0, the feature is disabled.

When the feature is enabled, the memory for the framebuffer is
allocated from the MM/IOMMU heap. The overlay play interface is
used to post the composited framebuffer on the display. The destination
rectangle for the framebuffer is set to be the panel resolution.

For bypass buffers, the destination rectangle is upscaled/downscaled as
per the panel resolution.

Change-Id: Ieff733950fc4945c875da803122d6b5add24692f