display : Compose front camera frames with MDP
Composing front camera preview frames with GPU is causing
high power consumption. This change will change the composition
of front camera frames to MDP and fixes inverted frames issue
during transition.

When flip and rotation(90 or 270) are present alter the flip value,
as GPU is doing the flip and rotation in opposite order to that of MDP3.0

Change-Id: I82bec10024904cad084d281dde56d502fdabf61e
CRs-Fixed: 353251