libhwcomposer:Use copybit if useCopybit is true
This change makes sure that in case of prepareOverlay failure
the composition type of both video and UI layers is the same.
This prevents the cases when video is assigned C2D and UI layers
are assigned GPU for composition which results in video not
being shown after composition.

CRs-Fixed: 400083

Change-Id: I65fc47c7f145326bb3f8406133365f47824b137a