hwc: Reset Layer cache on idle timeout
In the case where MDP composition is possible but we fall back to
the idle timeout, there is a possibility that all of the layer
handles that were cached in the _earlier_ idle timeout are the
same in the _next_ idle timeout. This could cause us to display a
stale frame. Reset LayerCache in prepare if we are in the idle
frame.

CRs-fixed: 484324
Change-Id: I3ab1b0f28044975978a32e082976b038242697d0