hwc: mdpcomp: split: Count total pipes needed per display also
When checking for pipes needed for each mixer, the pipes completely
unused are counted twice.
Add a per display check on top of per mixer check to make sure this
is taken care of.

For example: Each mixer needs 4 pipes, and total completely unused
pipes are 6. Each mixer will get 6 as available pipes, which is ok
at a mixer level, but at a display level 8 pipes are needed and 6
available. Need to account for that

Conflicts:

	liboverlay/overlay.h

Change-Id: I9811255aab96c7fe47331f8aa125fef2a4a2f704