hwc: vpuclient: vpuClient implementation
Vpuclient is the client side for libvpu in hwc. It follows the
similar pattern of prepare and draw. It has 2 step prepare
including: setVpuSession and prepare. The setVpuSession function
passes all the layers from the SF list to libvpu, which marks
the layers that it can support. After this function,
mdpcomp->prepare allocates the pipe(s) for the vpu layers, and
the allocated pipes are sent to libvpu in the vpuclient prepare
function. The vpuclient draw function passes the handle to the
libvpu

Change-Id: I6ed2f8f15f07a67b08c4f17f85714d8c005b15af