hwc: Ignore force overlay_set on mdp3 targets
Since mdp3 only has one dma pipe for overlay, which in the driver
is just a sw overlay, there is no need to force a overlay_set call
when HWC_GEOMETRY_CHANGED is set.
Also we are moving the blocking wait for destination buffer fence
to inside the blit call occurs inside work queue for mdp3 targets.

Change-Id: Id0aed8732dc49b1041d8174a5ff38703c8023abf