hwc: Disabling the caching in case of undescan.
Since the Caching is enabled, changed underscan parameters
takes time to reflect for the new surfaces drawn. This result
into the black screen for some duration. With this change the
bug is not seen.
If Over Scan Composition parameters are set, then compose the
current frame else make isFBComposed flag true to put it in cache.

Change-Id: I29fee036a0ec6e82235659ed1e3e8606cd5369da