hwc: vpuclient: Added dummy buffer implementation
For the VPU-supported video, MDP expects something to be queued
for the first frame. It is required so that MDP could process the
tunnelling further on. In order to do so, hwc queues a dummy
buffer filled with black color. For future, this needs to be
changed to SOLID_FILL implementation

Change-Id: I1f92bb8281f764315b4c42d87c6aad382d3418f9