hwc: Do not program MDP with invalid destination rectangle
In a frame, if a layer is hidden under equal opaque layer, a HAL
optimization results in deducting lower layer with upper, hence
results in a rectangle with zero height. So MDP complains about
invalid destination rectangle in OVERLAY_PREPARE and frame falls
back to GPU Composition. So avoid invalid MDP programming.

Change-Id: I54342866ad8c161ffce7d8047237ebbe06d31cf6