hwc: Modify MDPOnly composition to allow only secure RGB layers
MDPOnly composition needs to cater for the use case when we have
secure RGB layers that require MDP composition. For non-secure RGB
layers we can afford to send them for GPU composition. This is
especially needed for the following use case: full MDP composition
on primary (including YUV layers), followed by external connection.
For the mentioned use case, we need to fall back to Video and then
GPU composition in in order to reclaim pipe resources for external
usage.

Change-Id: I6da3ca48cf6e38e767fecfbfee6b06e82ba00a51