hwc: Complete DMA transition with a refresh
In case of secure videos, if DMA pipe transition is needed and if
there are no subsequent draw rounds, black content continues to be
shown. This change triggers a refresh in such cases to complete the
DMA pipe transition. It also allows composition to go through if the
previous and current frame are same with GPU composition, if secure
videos are present.

Change-Id: I5202f0f170ff5d1a0f8280fcf385a23749b5d0a4