hwc: Set minfps as part of timeout_handler during idlefallback
It is possible that mdss driver updates the refresh-rate back
to 60 during idletimeout when screen is touched, in anticipation
of updates. However, display-hal is agnostic of those updates.

It might be the case that touch sometimes may not result in any
legitimate updates to display. In such cases, mdss driver triggers
idle notify after certain timeout period. Reset the dyn-fps
back to min-fps supported by the panel as part of handling it.

Change-Id: I1195039db7c29d0bf60094794aca95651f1b3eec